Bug fixes in synthetic corpus generator.
Generated documents sometimes had one more term than max_terms. This was
because the title stream term was not included in the count of terms
generated. Sometimes the title term was the same as a body term,
resulting in the correct count. Other times the title term was distinct,
leading to a document with on term too many. The fix was to make the
title term encoding unique (it's "tID" where ID is the document id) and
then count it. This means that the minimum legal value for min_terms is
65 - one for the title, and one for each of the bits in the document id.

Seeded numpy.random so that runs are reproducible.

Fixed bug where terms were only getting generated for the 4 least
significant bits of the document id. Now all 64 bits are considered.